backend/rabbitMQ/requirements.txt
pika>=1.3.1
requests>=2.31.0
python-dotenv==1.0.1